+++ orchestration/src/test/java/com/sap/ai/sdk/orchestration/OrchestrationUnitTest.java
@@ -46,6 +46,7 @@ import com.sap.ai.sdk.orchestration.model.DPIEntities;
 import com.sap.ai.sdk.orchestration.model.DataRepositoryType;
 import com.sap.ai.sdk.orchestration.model.DocumentGroundingFilter;
 import com.sap.ai.sdk.orchestration.model.ErrorResponse;
+import com.sap.ai.sdk.orchestration.model.ErrorResponseError;
 import com.sap.ai.sdk.orchestration.model.FilteringStreamOptions;
 import com.sap.ai.sdk.orchestration.model.GenericModuleResult;
 import com.sap.ai.sdk.orchestration.model.GroundingFilterSearchConfiguration;
@@ -269,6 +270,28 @@ class OrchestrationUnitTest {
   }
 
   @Test
+  void testGroundingWithConvenience() throws IOException {
+    stubFor(
+        post(urlPathEqualTo("/v2/completion"))
+            .willReturn(
+                aResponse()
+                    .withBodyFile("groundingResponse.json")
+                    .withHeader("Content-Type", "application/json")));
+
+    var dbFilters = DocumentGroundingFilter.create().dataRepositoryType(DataRepositoryType.VECTOR);
+
+    var groundingConfig = Grounding.create().metadataParams("foo", "bar").filters(dbFilters);
+    var prompt = groundingConfig.createGroundingPrompt("Hello, what do you know?");
+    var configWithGrounding = config.withGrounding(groundingConfig);
+
+    final var response = client.chatCompletion(prompt, configWithGrounding);
+
+    final String request = fileLoaderStr.apply("groundingRequestMetadata.json");
+    verify(
+        postRequestedFor(urlPathEqualTo("/v2/completion")).withRequestBody(equalToJson(request)));
+  }
+
+  @Test
   void testGroundingWithHelpSapCom() throws IOException {
     stubFor(
         post(urlPathEqualTo("/v2/completion"))
@@ -396,13 +419,11 @@ class OrchestrationUnitTest {
                       "Request failed with status 400 (Bad Request): Missing required parameters: ['input']");
               assertThat(e.getErrorResponseStreaming()).isNull();
               assertThat(e.getErrorResponse()).isNotNull();
-              assertThat(e.getErrorResponse().getError().getMessage())
-                  .isEqualTo("Missing required parameters: ['input']");
-              assertThat(e.getErrorResponse().getError().getCode()).isEqualTo(SC_BAD_REQUEST);
-              assertThat(e.getErrorResponse().getError().getRequestId())
-                  .isEqualTo("51043a32-01f5-429a-b0e7-3a99432e43a4");
-              assertThat(e.getErrorResponse().getError().getLocation())
-                  .isEqualTo("Module: Templating");
+              val error = ((ErrorResponseError.InnerError) e.getErrorResponse().getError()).value();
+              assertThat(error.getMessage()).isEqualTo("Missing required parameters: ['input']");
+              assertThat(error.getCode()).isEqualTo(SC_BAD_REQUEST);
+              assertThat(error.getRequestId()).isEqualTo("51043a32-01f5-429a-b0e7-3a99432e43a4");
+              assertThat(error.getLocation()).isEqualTo("Module: Templating");
             });
   }
 
@@ -531,8 +552,9 @@ class OrchestrationUnitTest {
               final var errorResponse = e.getErrorResponse();
               assertThat(errorResponse).isNotNull();
               assertThat(errorResponse).isInstanceOf(ErrorResponse.class);
-              assertThat(errorResponse.getError().getCode()).isEqualTo(SC_BAD_REQUEST);
-              assertThat(errorResponse.getError().getMessage())
+              val error = ((ErrorResponseError.InnerError) errorResponse.getError()).value();
+              assertThat(error.getCode()).isEqualTo(SC_BAD_REQUEST);
+              assertThat(error.getMessage())
                   .isEqualTo(
                       "400 - Filtering Module - Input Filter: Prompt filtered due to safety violations. Please modify the prompt and try again.");